// src/data/generated/epicycleFits.ts
export const EPICYCLE_PLANET_IDS = [
  'moon',
  'mercury',
  'venus',
  'sun',
  'mars',
  'jupiter',
  'saturn',
] as const